requirements.txt
# ------
# Basic
# ------
pip>=9

# ------
# GCP
# ------
google-cloud-storage
